+++ sdk/core/azure-core/tests/async_tests/test_rest_polling_async.py
@@ -143,3 +143,23 @@ async def test_continuation_token(client, lro_poller, deserialization_callback):
     )
     result = await new_poller.result()
     assert result == {"location_result": True}
+
+
+@pytest.mark.asyncio
+async def test_put_location_api_version_add(lro_poller):
+    result = await (
+        await lro_poller(
+            HttpRequest("PUT", "/polling/location/api-version/add"), lro_options={"api_version": "2023-11-15"}
+        )
+    ).result()
+    assert result["location_result"]
+
+
+@pytest.mark.asyncio
+async def test_put_location_api_version_replace(lro_poller):
+    result = await (
+        await lro_poller(
+            HttpRequest("PUT", "/polling/location/api-version/replace"), lro_options={"api_version": "2023-11-15"}
+        )
+    ).result()
+    assert result["location_result"]