+++ sdk/core/azure-core/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### Other Changes
 
+- Support add or replace api-version when lro polling #33192
+
 ## 1.29.5 (2023-10-19)
 
 ### Bugs Fixed